Add profile section controller parent class
